requirements.txt
# MNIST手写数字识别项目依赖

# 核心深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 数据处理和科学计算
numpy>=1.19.0
Pillow>=8.0.0

# 机器学习评估
scikit-learn>=0.24.0

# 数据可视化
matplotlib>=3.3.0
seaborn>=0.11.0

# 可选依赖（用于更高级的功能）
# ipython>=7.0.0
# jupyter>=1.0.0
# tensorboard>=2.4.0
# Web框架
flask>=2.0.0
